finishes transition to long doubles

+++ src/Admixture.cpp
@@ -1,7 +1,7 @@
 /*
  * Authors: Gustavo V. Barroso
  * Created: 21/04/2023
- * Last modified: 23/08/2023
+ * Last modified: 06/06/2023
  *
  */
 
@@ -19,7 +19,7 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
   {
     size_t ancFromId = 0;
     size_t ancToId = 0;
-    double f = -1.;
+    long double f = -1.;
 
     for(size_t j = 0; j < numPops; ++j)
     {
@@ -33,7 +33,7 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
 
     if(f > 0.)
     {
-      std::vector<Eigen::Triplet<double>> coeffs(0);
+      std::vector<Eigen::Triplet<long double>> coeffs(0);
       coeffs.reserve(3 * sizeOfBasis);
 
       for(auto it = std::begin(sslib.getMoments()); it != std::end(sslib.getMoments()); ++it)
@@ -51,11 +51,11 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
             {
               col = sslib.findCompressedIndex(*it2nd);
 
-              double mig = (*it)->countInstances(ancToId) - (*it2nd)->countInstances(ancToId);
-              double nat = (*it2nd)->countInstances(ancToId);
-              double y = std::pow((1. - f), nat) * std::pow(f, mig) / ((*it)->getNumberOfAliases() + 1);
+              long double mig = (*it)->countInstances(ancToId) - (*it2nd)->countInstances(ancToId);
+              long double nat = (*it2nd)->countInstances(ancToId);
+              long double y = std::pow((1. - f), nat) * std::pow(f, mig) / ((*it)->getNumberOfAliases() + 1);
 
-              coeffs.emplace_back(Eigen::Triplet<double>(row, col, y));
+              coeffs.emplace_back(Eigen::Triplet<long double>(row, col, y));
             }
           }
 
@@ -81,14 +81,14 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
                 size_t t = 1 + (*it)->countInstances(ancToId);
                 size_t a = 1 + ((*itCmp)->getPopIndices()[0] == ancToId);
                 size_t b = t - a;
-                double c = std::pow(-1., (*itCmp)->countInstances(ancFromId));
+                long double c = std::pow(-1., (*itCmp)->countInstances(ancFromId));
 
                 if((*itCmp)->getPopIndices()[0] == ancFromId)
                   c *= -1.;
 
-                double x = (0.25 * (*it)->countInstances(ancToId) * c * std::pow((1. - f), a) * std::pow(f, b)) / ((*it)->getNumberOfAliases() + 1);
+                long double x = (0.25 * (*it)->countInstances(ancToId) * c * std::pow((1. - f), a) * std::pow(f, b)) / ((*it)->getNumberOfAliases() + 1);
 
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, x));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, x));
               }
 
               // contributions from pi2 moments
@@ -96,9 +96,9 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
               {
                 col = sslib.findCompressedIndex(*itCmp);
                 size_t parentPopIdCount =(*itCmp)->countInstances(ancFromId);
-                double x = std::pow((1. - f), 2.) * std::pow(f, 2.);
-                double y = std::pow(-1., parentPopIdCount) * x;
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, y));
+                long double x = std::pow((1. - f), 2.) * std::pow(f, 2.);
+                long double y = std::pow(-1., parentPopIdCount) * x;
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, y));
               }
             }
           }
@@ -114,86 +114,86 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
               if((*it)->countInstances(ancToId) == 3)
               {
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 4. * std::pow(f, 3.) * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 4. * std::pow(f, 3.) * (1. - f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancFromId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancToId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 2. * std::pow(f, 2.) * (1. - f) * (1. - 2 * f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancToId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -4. * std::pow(f, 2.) * std::pow(1. - f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -4. * std::pow(f, 2.) * std::pow(1. - f, 2.)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancToId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -4. * std::pow(f, 2.) * std::pow(1. - f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -4. * std::pow(f, 2.) * std::pow(1. - f, 2.)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancFromId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancToId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancFromId, ancToId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f) * std::pow(1. - 2. * f, 2.)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancToId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancToId, ancToId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancToId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancFromId, ancToId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * std::pow(1. - f, 2.) * (1. - 2. * f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancToId, ancToId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 4. * f * std::pow(1. - f, 3.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 4. * f * std::pow(1. - f, 3.)));
               }
 
               // case 2: Dr TFT / TTF NOTE "forced" through some differences with the "naive" Mathematica notebook to match moments.LD admix matrix
               else if(((*it)->getPopIndices()[1] == ancToId && (*it)->getPopIndices()[2] == ancFromId) || ((*it)->getPopIndices()[1] == ancFromId && (*it)->getPopIndices()[2] == ancToId))
               {
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 2. * std::pow(f, 2.) * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 2. * std::pow(f, 2.) * (1. - f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 2. * f * (1. - f) * (1. - 3. * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 2. * f * (1. - f) * (1. - 3. * f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancToId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * std::pow(1. - f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * std::pow(1. - f, 2.)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * (1. - f) * (1. - 2. * f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * (1. - f) * (1. - 2. * f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancToId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 2. * f * std::pow(1. - f, 2.)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 2. * f * std::pow(1. - f, 2.)));
               }
 
               // case 3: Dr TFF
               else if((*it)->getPopIndices()[1] == ancFromId && (*it)->getPopIndices()[2] == ancFromId)
               {
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, 4. * f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, 4. * f * (1. - f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * (1. - f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancFromId, ancFromId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * (1. - f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancToId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * (1. - f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancFromId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, -2. * f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, -2. * f * (1. - f)));
 
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancFromId, ancToId, ancToId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancFromId, ancFromId, ancToId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f)));
                 col = sslib.findCompressedIndex(sslib.findPi2Index(ancToId, ancFromId, ancToId, ancFromId, p));
-                coeffs.emplace_back(Eigen::Triplet<double>(row, col, f * (1. - f)));
+                coeffs.emplace_back(Eigen::Triplet<long double>(row, col, f * (1. - f)));
               }
 
               // case 4: Dr TTx / TxT
@@ -206,11 +206,11 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
                     auto tmp = std::dynamic_pointer_cast<Pi2Moment>(*itCmp);
                     col = sslib.findCompressedIndex(tmp);
 
-                    double c = std::pow(-1., tmp->countInstances(ancToId) + 1);
-                    double z = (tmp->getLeftHetStat()->isCrossPop() && tmp->getRightHetStat()->isCrossPop());
-                    double x = 1 + (tmp->getLeftHetStat()->countInstances(ancFromId) == 2 || tmp->getRightHetStat()->countInstances(ancFromId) == 2);
-                    double y = 3. - z - x;
-                    double w = std::pow(f, x) * std::pow((1. - f), y) * std::pow(1. - 2 * f, z);
+                    long double c = std::pow(-1., tmp->countInstances(ancToId) + 1);
+                    long double z = (tmp->getLeftHetStat()->isCrossPop() && tmp->getRightHetStat()->isCrossPop());
+                    long double x = 1 + (tmp->getLeftHetStat()->countInstances(ancFromId) == 2 || tmp->getRightHetStat()->countInstances(ancFromId) == 2);
+                    long double y = 3. - z - x;
+                    long double w = std::pow(f, x) * std::pow((1. - f), y) * std::pow(1. - 2 * f, z);
 
                     if(z == 0)
                     {
@@ -219,7 +219,7 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
                         w *= -1;
                     }
 
-                    coeffs.emplace_back(Eigen::Triplet<double>(row, col, c * w));
+                    coeffs.emplace_back(Eigen::Triplet<long double>(row, col, c * w));
                   }
                 }
               }
@@ -233,8 +233,8 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
                   {
                     auto tmp = std::dynamic_pointer_cast<Pi2Moment>(*itCmp);
                     col = sslib.findCompressedIndex(tmp);
-                    double c = std::pow(-1., tmp->countInstances(ancToId));
-                    coeffs.emplace_back(Eigen::Triplet<double>(row, col, c * 2 * f * (1. - f)));
+                    long double c = std::pow(-1., tmp->countInstances(ancToId));
+                    coeffs.emplace_back(Eigen::Triplet<long double>(row, col, c * 2 * f * (1. - f)));
                   }
                 }
               }
@@ -248,9 +248,9 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
                   {
                     auto tmp = std::dynamic_pointer_cast<Pi2Moment>(*itCmp);
                     col = sslib.findCompressedIndex(tmp);
-                    double c = std::pow(-1., tmp->countInstances(ancToId));
-                    double d = 2. * (1. + ((*it)->getPopIndices()[1] == (*it)->getPopIndices()[2]));
-                    coeffs.emplace_back(Eigen::Triplet<double>(row, col, c * d * f * (1. - f)));
+                    long double c = std::pow(-1., tmp->countInstances(ancToId));
+                    long double d = 2. * (1. + ((*it)->getPopIndices()[1] == (*it)->getPopIndices()[2]));
+                    coeffs.emplace_back(Eigen::Triplet<long double>(row, col, c * d * f * (1. - f)));
                   }
                 }
               }
@@ -259,7 +259,7 @@ void Admixture::setUpMatrices_(const SumStatsLibrary& sslib)
         }
       } // ends loop over moments
 
-      Eigen::SparseMatrix<double> mat(sizeOfBasis, sizeOfBasis);
+      Eigen::SparseMatrix<long double> mat(sizeOfBasis, sizeOfBasis);
       mat.setFromTriplets(std::begin(coeffs), std::end(coeffs));
       mat.makeCompressed();
       matrices_.emplace_back(mat);
@@ -290,14 +290,14 @@ void Admixture::updateMatrices_()
 
           if(hasParameter(paramName))
           {
-            double newVal = getParameterValue(paramName);
+            long double newVal = getParameterValue(paramName);
             littleAdmixMat_(i, j) = newVal;
           }
 
           else
           {
             paramName = "a_" + bpp::TextTools::toString(jd) + "_" + bpp::TextTools::toString(id); // switch order
-            double newVal = getParameterValue(paramName);
+            long double newVal = getParameterValue(paramName);
             littleAdmixMat_(i, j) = 1. - newVal;
           }
         }
@@ -324,7 +324,7 @@ void Admixture::assembleTransitionMatrix_()
   // only adds 1 to main diagonal of empty rows
   for(int i = 0; i < transition_.rows(); ++i)
   {
-    double rowSum = 0.;
+    long double rowSum = 0.;
 
     for(int j = 0; j < transition_.cols(); ++j)
       rowSum += transition_.coeffRef(i, j);
