+++ src/Recombination.cpp
@@ -1,7 +1,7 @@
 /*
  * Authors: Gustavo V. Barroso
  * Created: 09/08/2022
- * Last modified: 30/05/2024
+ * Last modified: 06/06/2024
  *
  */
 
@@ -65,8 +65,8 @@ void Recombination::updateMatrices_()
     size_t id = popIndices_[i];
     std::string paramName = "r_" + bpp::TextTools::toString(id);
 
-    double prevVal = prevParams_.getParameterValue(paramName);
-    double newVal = getParameterValue(paramName);
+    long double prevVal = prevParams_.getParameterValue(paramName);
+    long double newVal = getParameterValue(paramName);
 
     if(newVal != prevVal)
       matrices_[i] *= (newVal / prevVal);
